Add files via upload
The "DDoS Attack Prediction" project is dedicated to forecasting and mitigating Distributed Denial of Service (DDoS) attacks on computer networks and online services. This initiative involves data analysis, pattern recognition, and machine learning techniques to identify early warning signs of potential DDoS attacks. By predicting and proactively responding to these threats, the project aims to ensure the availability and resilience of digital services, prevent disruptions, and maintain the integrity of network infrastructure. Through continuous monitoring and predictive modeling, it bolsters the defense against one of the most prevalent and disruptive forms of cyberattacks.