Add files via upload
The "IDS for Cloud Security" project focuses on the development and implementation of Intrusion Detection Systems (IDS) tailored for cloud computing environments. It entails the design of sophisticated algorithms and monitoring mechanisms to detect and respond to security threats within cloud infrastructure. This project addresses the unique challenges posed by the dynamic and scalable nature of cloud platforms, aiming to safeguard data and applications hosted in the cloud. By enhancing threat detection, real-time alerting, and incident response in the cloud, the project bolsters the security and reliability of cloud services, ensuring the confidentiality and integrity of critical data.